chore: no more mobile exports
