+++ frontend/src/scenes/session-recordings/player/player-meta/PlayerMetaLinks.tsx
@@ -1,7 +1,6 @@
 import { IconDownload, IconEllipsis, IconMinusSmall, IconNotebook, IconPlusSmall, IconTrash } from '@posthog/icons'
 import { LemonButton, LemonButtonProps, LemonDialog, LemonMenu, LemonMenuItems } from '@posthog/lemon-ui'
 import { useActions, useValues } from 'kea'
-import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { IconComment } from 'lib/lemon-ui/icons'
 import { useMemo } from 'react'
 import { useNotebookNode } from 'scenes/notebooks/Nodes/NotebookNodeContext'
@@ -32,7 +31,7 @@ function PinToPlaylistButton(): JSX.Element {
             size="xsmall"
             onClick={() => {
                 if (nodeLogic) {
-                    // If we are in a node, then pinning should persist the recording
+                    // If we are in a node, then pinning should persist that recording
                     maybePersistRecording()
                 }
 
@@ -142,8 +141,6 @@ const MenuActions = ({ size }: { size: PlayerMetaBreakpoints }): JSX.Element => 
     const { logicProps } = useValues(sessionRecordingPlayerLogic)
     const { deleteRecording, setIsFullScreen, exportRecordingToFile } = useActions(sessionRecordingPlayerLogic)
 
-    const hasMobileExportFlag = useFeatureFlag('SESSION_REPLAY_EXPORT_MOBILE_DATA')
-    const hasMobileExport = window.IMPERSONATED_SESSION || hasMobileExportFlag
     const isStandardMode =
         (logicProps.mode ?? SessionRecordingPlayerMode.Standard) === SessionRecordingPlayerMode.Standard
 
@@ -172,7 +169,7 @@ const MenuActions = ({ size }: { size: PlayerMetaBreakpoints }): JSX.Element => 
                 label: '.json',
                 status: 'default',
                 icon: <IconDownload />,
-                onClick: () => exportRecordingToFile(false),
+                onClick: () => exportRecordingToFile(),
                 tooltip: 'Export recording to a JSON file. This can be loaded later into PostHog for playback.',
             },
         ]
@@ -180,17 +177,6 @@ const MenuActions = ({ size }: { size: PlayerMetaBreakpoints }): JSX.Element => 
             itemsArray.unshift({
                 label: () => <AddToNotebookButton fullWidth={true} />,
             })
-        }
-        if (hasMobileExport) {
-            isStandardMode &&
-                itemsArray.push({
-                    label: 'DEBUG - mobile.json',
-                    status: 'default',
-                    icon: <IconDownload />,
-                    onClick: () => exportRecordingToFile(true),
-                    tooltip:
-                        'DEBUG - ONLY VISIBLE TO POSTHOG STAFF - Export untransformed recording to a file. This can be loaded later into PostHog for playback.',
-                })
         }
         if (logicProps.playerKey !== 'modal') {
             isStandardMode &&
@@ -202,7 +188,7 @@ const MenuActions = ({ size }: { size: PlayerMetaBreakpoints }): JSX.Element => 
                 })
         }
         return itemsArray
-    }, [logicProps.playerKey, onDelete, exportRecordingToFile, hasMobileExport, size])
+    }, [logicProps.playerKey, onDelete, exportRecordingToFile, size])
 
     return (
         <LemonMenu items={items} buttonSize="xsmall">
