+++ frontend/src/scenes/session-recordings/player/sessionRecordingDataLogic.ts
@@ -41,12 +41,13 @@ import { decompressEvent } from './snapshot-processing/decompress'
 import { deduplicateSnapshots } from './snapshot-processing/deduplicate-snapshots'
 import {
     getHrefFromSnapshot,
+    patchMetaEventIntoMobileData,
     patchMetaEventIntoWebData,
     ViewportResolution,
 } from './snapshot-processing/patch-meta-event'
-import { patchMetaEventIntoMobileData } from './snapshot-processing/patch-meta-event'
 import { throttleCapture } from './snapshot-processing/throttle-capturing'
 import { createSegments, mapSnapshotsToWindowId } from './utils/segmenter'
+
 const IS_TEST_MODE = process.env.NODE_ENV === 'test'
 const TWENTY_FOUR_HOURS_IN_MS = 24 * 60 * 60 * 1000 // +- before and after start and end of a recording to query for session linked events.
 const FIVE_MINUTES_IN_MS = 5 * 60 * 1000 // +- before and after start and end of a recording to query for events related by person.
@@ -71,18 +72,15 @@ function hasAnyWireframes(snapshotData: Record<string, any>[]): boolean {
  *
  * If it can't be case as eventWithTime by this point then it's probably not a valid event anyway
  */
-function coerceToEventWithTime(d: unknown, withMobileTransformer: boolean, sessionRecordingId: string): eventWithTime {
-    // we decompress first so that we could support partial compression on mobile in future
+function coerceToEventWithTime(d: unknown, sessionRecordingId: string): eventWithTime {
+    // we decompress first so that we could support partial compression on mobile in the future
     const currentEvent = decompressEvent(d, sessionRecordingId)
-    return withMobileTransformer
-        ? postHogEEModule?.mobileReplay?.transformEventToWeb(currentEvent) || (currentEvent as eventWithTime)
-        : (currentEvent as eventWithTime)
+    return postHogEEModule?.mobileReplay?.transformEventToWeb(currentEvent) || (currentEvent as eventWithTime)
 }
 
 export const parseEncodedSnapshots = async (
     items: (RecordingSnapshot | EncodedRecordingSnapshot | string)[],
-    sessionId: string,
-    withMobileTransformer: boolean = true
+    sessionId: string
 ): Promise<RecordingSnapshot[]> => {
     if (!postHogEEModule) {
         postHogEEModule = await posthogEE()
@@ -126,7 +124,7 @@ export const parseEncodedSnapshots = async (
             }
 
             return snapshotData.flatMap((d: unknown) => {
-                const snap = coerceToEventWithTime(d, withMobileTransformer, sessionId)
+                const snap = coerceToEventWithTime(d, sessionId)
 
                 const baseSnapshot: RecordingSnapshot = {
                     windowId: snapshotLine['window_id'] || snapshotLine['windowId'],
@@ -169,22 +167,9 @@ export interface SessionRecordingDataLogicProps {
 
 async function processEncodedResponse(
     encodedResponse: (EncodedRecordingSnapshot | string)[],
-    props: SessionRecordingDataLogicProps,
-    featureFlags: FeatureFlagsSet
-): Promise<{ transformed: RecordingSnapshot[]; untransformed: RecordingSnapshot[] | null }> {
-    let untransformed: RecordingSnapshot[] | null = null
-
-    const transformed = await parseEncodedSnapshots(encodedResponse, props.sessionRecordingId)
-
-    if (featureFlags[FEATURE_FLAGS.SESSION_REPLAY_EXPORT_MOBILE_DATA]) {
-        untransformed = await parseEncodedSnapshots(
-            encodedResponse,
-            props.sessionRecordingId,
-            false // don't transform mobile data
-        )
-    }
-
-    return { transformed, untransformed }
+    props: SessionRecordingDataLogicProps
+): Promise<RecordingSnapshot[] | null> {
+    return await parseEncodedSnapshots(encodedResponse, props.sessionRecordingId)
 }
 
 export type SourceKey = `${SnapshotSourceType}-${string}`
@@ -365,13 +350,9 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
                         throw e
                     })
 
-                    const { transformed, untransformed } = await processEncodedResponse(
-                        response,
-                        props,
-                        values.featureFlags
-                    )
+                    const transformed = await processEncodedResponse(response, props)
 
-                    return { snapshots: transformed, untransformed_snapshots: untransformed ?? undefined, source }
+                    return { snapshots: transformed ?? undefined, source }
                 },
             },
         ],
@@ -913,19 +894,6 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
             },
         ],
 
-        untransformedSnapshots: [
-            (s) => [s.snapshotSources, s.snapshotsBySource],
-            (sources, snapshotsBySource): RecordingSnapshot[] => {
-                const allSnapshots =
-                    sources?.flatMap((source) => {
-                        const sourceKey = getSourceKey(source)
-                        return snapshotsBySource?.[sourceKey]?.untransformed_snapshots || []
-                    }) ?? []
-
-                return deduplicateSnapshots(allSnapshots)
-            },
-        ],
-
         snapshotsByWindowId: [
             (s) => [s.snapshots],
             (snapshots): Record<string, eventWithTime[]> => {
@@ -1005,18 +973,14 @@ export const sessionRecordingDataLogic = kea<sessionRecordingDataLogicType>([
         ],
 
         createExportJSON: [
-            (s) => [s.sessionPlayerMetaData, s.snapshots, s.untransformedSnapshots],
-            (
-                sessionPlayerMetaData,
-                snapshots,
-                untransformedSnapshots
-            ): ((exportUntransformedMobileSnapshotData: boolean) => ExportedSessionRecordingFileV2) => {
-                return (exportUntransformedMobileSnapshotData: boolean) => ({
+            (s) => [s.sessionPlayerMetaData, s.snapshots],
+            (sessionPlayerMetaData, snapshots): (() => ExportedSessionRecordingFileV2) => {
+                return () => ({
                     version: '2023-04-28',
                     data: {
                         id: sessionPlayerMetaData?.id ?? '',
                         person: sessionPlayerMetaData?.person,
-                        snapshots: exportUntransformedMobileSnapshotData ? untransformedSnapshots : snapshots,
+                        snapshots: snapshots,
                     },
                 })
             },
